pyproject.toml
[project]
name = "training-langchainagenticudemy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "black>=26.3.1",
    "chromadb>=1.5.5",
    "grandalf>=0.8",
    "httpx>=0.28.1",
    "isort>=8.0.1",
    "langchain>=1.2.13",
    "langchain-anthropic>=1.4.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.12",
    "langchain-pinecone>=0.2.13",
    "langchain-tavily>=0.2.17",
    "langchainhub>=0.1.21",
    "langgraph>=1.1.3",
    "markdownify>=1.2.2",
    "mcp[cli]>=1.27.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
    "pytest-socket>=0.7.0",
    "pytest-timeout>=2.4.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "requests>=2.33.0",
    "ruff>=0.15.10",
    "tavily-python>=0.7.23",
]
